Parse obj's values into floats, not doubles, which should hopefully speed things up a tid bit

+++ d3d12/ObjFileLoader.h
@@ -106,8 +106,8 @@ public:
   };
 
   struct AxisAlignedBounds {
-    double max[3];
-    double min[3];
+    float max[3];
+    float min[3];
   };
 
   std::vector<Vertex> m_vertices;
